Update sourcecred_run so paths aren't hardcoded
You must now set $SC_REPO_DIR and run the script from the root of this
repository for it to work. It fails cleanly if either is not true.

+++ sourcecred_run.sh
@@ -2,15 +2,36 @@
 
 set -eu
 
-DEMO_DIR="/Users/sben/Documents/coins/sourcecred/demos/makerdao_trial/sourcecred"
-SOURCECRED_DIR="/Users/sben/Documents/coins/sourcecred/sourcecred"
-SOURCECRED_CLI="/Users/sben/Documents/coins/sourcecred/sourcecred/bin/sourcecred.js"
-export SOURCECRED_DIRECTORY="$DEMO_DIR/sourcecred_data"
-SITE_DIR="$DEMO_DIR/site"
+# requires that the user has set a SC_REPO_DIR environment variable
+# and it points to a repository containing sourcecred, where `yarn backend`
+# has been run to build the CLI.
+# requires that this script be run from the root of the makerdao-cred repo
 
-node "$SOURCECRED_CLI" discourse https://forum.makerdao.com
+main() {
+  DEMO_DIR=$(pwd)
+  SITE_DIR="$DEMO_DIR/docs"
+  if ! [ -d "${SITE_DIR}" ]; then
+    die "Can't find site dir; probably running script from wrong location"
+  fi
 
-(cd "$SOURCECRED_DIR" && yarn build --output-path "$SITE_DIR")
-mkdir -p "$SITE_DIR/api/v1"
-cp -r "$SOURCECRED_DIRECTORY" "$SITE_DIR/api/v1/data"
-rm -rf "$SITE_DIR/api/v1/data/cache"
+  SOURCECRED_CLI="$SC_REPO_DIR/bin/sourcecred.js"
+  if ! [ -f "${SOURCECRED_CLI}" ]; then
+    die "Can't find sourcecred CLI"
+  fi
+  export SOURCECRED_DIRECTORY="$DEMO_DIR/sourcecred_data"
+
+  node "$SOURCECRED_CLI" discourse https://forum.makerdao.com
+
+  (cd "$SC_REPO_DIR" && yarn build --output-path "$SITE_DIR")
+  mkdir -p "$SITE_DIR/api/v1"
+  cp -r "$SOURCECRED_DIRECTORY" "$SITE_DIR/api/v1/data"
+  rm -rf "$SITE_DIR/api/v1/data/cache"
+  cp "$DEMO_DIR/CNAME" "$SITE_DIR"
+}
+
+die() {
+    printf >&2 'fatal: %s\n' "$@"
+    exit 1
+}
+
+main
